Commit after: 1. User modes 2. Decoration of Edit page 3. Decoration in general

+++ react-client/src/components/common/UserMenu/UserMenu.js
@@ -1,16 +1,37 @@
-import React from 'react'
+import React, {useEffect, useState} from 'react'
 import style from './UserMenu.module.scss'
 import {Link, useNavigate} from 'react-router-dom'
 
 const UserMenu = () => {
   const navigate = useNavigate()
+  const [username, setUsername] = useState('')
+  const [logStatus, setLogStatus] = useState('')
 
   const logout = () => {
-    localStorage.clear()
-    navigate('/login')
+    if (logStatus === 'Logout') {
+      localStorage.clear()
+      getUserName()
+      navigate('/podcast')
+    } else {
+      navigate('/admin')
+    }
   }
 
-  const username = localStorage.getItem('username')
+  const getUserName = () => {
+    const user = localStorage.getItem('username')
+    if (!user) {
+      setUsername('Menu')
+      setLogStatus('Login')
+    } else {
+      setUsername(user, setUsername)
+      setLogStatus('Logout')
+    }
+  }
+
+  useEffect(() => {
+    getUserName()
+  }, [])
+
   return (
     <div className={style.drop_down}>
       <button>
@@ -18,7 +39,7 @@ const UserMenu = () => {
       </button>
       <div className={style.list}>
         <div className={style.item} onClick={logout}>
-          logout
+          {logStatus}
         </div>
         <Link to="/podcast">
           <div className={style.item}>My Pod-Space</div>
